Handle `stylesheet` in `edoc_opts`
Instead of patching the HTML files, the plugin creates its own
stylesheet and passes it to EDoc using the `stylesheet` option.
`github-markdown.css` and `prism.css` are imported from that generated
CSS file.

If the user specified his own stylesheet, we can now import it from that
generated CSS file. This allows the user to provide his own CSS rules
and even override the other stylesheets we import.