Allow doclet and layout EDoc modules chaining
To implement this plugin, we override the `doclet` and `layout` modules
with our own. However, a user may want to set his own modules too.

With this patch, the plugin remembers the user-configured modules, if
any, and calls them or the default modules, before it patches the
output.

This is useful when a user wants to enable doc chunks (EEP-48) to be
able to get the API documentation from the Erlang shell or an IDE. For
that, he has to set `edoc_doclet_chunks` and `edoc_layout_chunks` in the
EDoc options. Thanks to this patch, the doc chunks are generated as well
as the improved HTML documentation.